Split point style builder from point style
